some fixes

+++ main.cpp
@@ -7,7 +7,7 @@ using namespace sf;
 using namespace std;
 void Draw(RenderWindow& window, Snake& snake, const float segmentSize) {
 	vector<vector<int>> state = snake.State();
-	const float scoreMargin = 2;
+	const int scoreMargin = 2;
 	window.clear();
 	for (size_t i = 0; i < state.size(); ++i) {
 		for (size_t j = 0; j < state[0].size(); ++j) {
@@ -32,7 +32,7 @@ void Draw(RenderWindow& window, Snake& snake, const float segmentSize) {
 	window.draw(interface);
 	Font font;
 	font.loadFromFile("Font.ttf");
-	Text score("", font, segmentSize * scoreMargin - 10);
+	Text score("", font, (unsigned int)segmentSize * scoreMargin - 10);
 	score.setString("Score: " + to_string(snake.Score()));
 	score.setFillColor(Color::White);
 	score.setPosition(Vector2f(10, 0));
